+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Add ability to override the runtime version per instance
 - Add in a unified modpack search to search across multiple platforms at once [#720]
 - Add category filtering when adding mods/shaders/worlds/resource packs
+- List the timeout and concurrent connection settings when downloads fail
 
 ### Fixes
 - Remove old OmitStackTraceInFastThrow JVM arg